Inject auth_check into lambda_handler as a kwarg
I'm not super pleased with the framework for this, but it's the
most-simple solution until I find a different abstraction to work with.

I could speculate a bunch of better ways to do this (wrappers,
call-stacks, "Apps", etc.) but this way I get to move my authentication
tests to the right module and get to wait until I have better view of
the abstraction.